Move Path to mod.rs of the path module
